Fixes. Account management icons.
- Removed language selection again (not ready for update)
- Fixed error when missing image, but tries to check if it's valid.
- Added icons to Login, Add new and Save account buttons.
- Fixed missing default colors.

+++ TcNo-Acc-Switcher-Server/Pages/General/GeneralFuncs.cs
@@ -131,7 +131,7 @@ namespace TcNo_Acc_Switcher_Server.Pages.General
             Globals.DebugWriteLine($@"[Func:General\GeneralFuncs.DeletedInvalidImage] filename={filename.Substring(filename.Length - 8, 8)}");
             try
             {
-                if (!IsValidGdiPlusImage(filename)) // Delete image if is not as valid, working image.
+	            if (File.Exists(filename) && !IsValidGdiPlusImage(filename)) // Delete image if is not as valid, working image.
                 {
                     File.Delete(filename);
                     return true;
